+++ Script/global.d.ts
@@ -17,23 +17,38 @@ import { Thenable } from "./thenable.d.ts";
 
 declare global {
     // -----------------------  API ---------------------------
+    interface ScriptPickerOption {
+        label: string
+        fn(): void
+        priority?: number
+        busy?: boolean
+        detail?: string
+        alwaysShow?: boolean
+        description?: string
+    }
     var Context: Promise<ExtensionContext>;
-    var Script: {
+    namespace Script {
         /**
          * Listen onActivate.
          * You can also use it to get `ExtensionContext`
          */
-        onActivate: (fn: (ctx: ExtensionContext) => any) => Promise<any>,
+        function onActivate(fn: (ctx: ExtensionContext) => any): Promise<any>
         /** This Event run when the Extention is deactivated */
-        onDeactivate: (fn: () => void) => void,
+        function onDeactivate(fn: () => void): void
         /** `show`, `hide` and `clear` ourput programmatically */
-        output: (method: 'show' | 'hide' | 'clear') => void
+        function output(method: 'show' | 'hide' | 'clear'): void;
+
+        /** @returns Cleanup function */
+        function picker(label: string, fn: () => void, priority = 0): () => void;
+        /** @returns Cleanup function */
+        function picker(scriptPickerOption: ScriptPickerOption): () => void;
+        function picker(any: ScriptPickerOption | string, fn: () => void, priority = 0): () => void;
     }
     /** Print message to `output` */
     function print(msg?: any): void;
     /** Print to `Output` with newline. */
     function println(msg?: any): void;
-    //---------------------------------------------------------
+    //---------------------------------------------------
 
     /**
      * The version of the editor.